开始搭建状态系统 版本1.2.3

+++ app/src/main/java/com/example/lifegame/data/entity/AttributeEntity.kt
@@ -8,8 +8,8 @@ data class AttributeEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0,
     val name: String,
-    val currentValue: Int,
-    val initialValue: Int,
+    val currentValue: Float,
+    val initialValue: Float,
     val colorHex: String,
     val sortOrder: Int = 0
 )
